Refactor to radiusProxyFromKubernetes it's own file.
