+++ go/chisel/share/tunnel/tunnel.go
@@ -3,20 +3,15 @@ package tunnel
 import (
 	"bytes"
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"errors"
-	"fmt"
 	"io/ioutil"
 	"log"
 	"net"
-	"net/http"
 	"os"
 	"sync"
 	"time"
 
 	"github.com/armon/go-socks5"
-	"github.com/inverse-inc/go-utils/sharedutils"
 	"github.com/inverse-inc/packetfence/go/chisel/share/cio"
 	"github.com/inverse-inc/packetfence/go/chisel/share/cnet"
 	"github.com/inverse-inc/packetfence/go/chisel/share/settings"
@@ -25,8 +20,6 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
 )
 
 // Config a Tunnel
@@ -112,55 +105,6 @@ func isPodReady(pod *v1.Pod) bool {
 	}
 
 	return false
-}
-
-const radiusAuthK8Filter = "app=radiusd-auth"
-
-func clientSetFromEnv() (*kubernetes.Clientset, error) {
-	host := os.Getenv("K8S_MASTER_URI")
-	if host == "" {
-		return nil, errors.New("K8_MASTER_URI is not defined")
-	}
-
-	token := os.Getenv("K8S_MASTER_TOKEN")
-	if token == "" {
-		return nil, errors.New("K8_MASTER_TOKEN is not defined")
-	}
-
-	return kubernetes.NewForConfigAndClient(
-		&rest.Config{
-			Host:        host,
-			BearerToken: token,
-		},
-		&http.Client{
-			Transport: &http.Transport{
-				TLSClientConfig: TLSConfigFromEnv(),
-			},
-		},
-	)
-}
-
-func TLSConfigFromEnv_() rest.TLSClientConfig {
-	caFile := sharedutils.EnvOrDefault("K8S_MASTER_CA_FILE", "/var/run/secrets/kubernetes.io/serviceaccount/ca.crt")
-	return rest.TLSClientConfig{
-		CAFile: caFile,
-	}
-}
-
-func TLSConfigFromEnv() *tls.Config {
-	caCerts := []byte(sharedutils.ReadFromFileOrStr(sharedutils.EnvOrDefault("KUBERNETES_CA_PATH", "/var/run/secrets/kubernetes.io/serviceaccount/ca.crt")))
-	rootCAs, _ := x509.SystemCertPool()
-	if rootCAs == nil {
-		rootCAs = x509.NewCertPool()
-	}
-
-	if ok := rootCAs.AppendCertsFromPEM(caCerts); !ok {
-		fmt.Println("No K8S CA cert appended, using system certs only")
-	}
-
-	return &tls.Config{
-		RootCAs: rootCAs,
-	}
 }
 
 // BindSSH provides an active SSH for use for tunnelling
